Add Studio Brussel streams
As published at https://www.vrt.be/nl/aanbod/kijk-en-luister/radio-luisteren/streamingslinks-radio/ by the Vlaamse Radio- en Televisieomroeporganisatie (Flemish Radio and Television Broadcasting Organization), or VRT, which is the national public-service broadcaster for the Flemish Region and Community of Belgium. The publication page explicitly mentions that the streams are free to use for non-commercial usage.